+++ rootfs/etc/s6-overlay/s6-rc.d/dnsmasq-init/init.sh
@@ -1,6 +1,7 @@
 #!/usr/bin/with-contenv bashio
+
 # ==============================================================================
-# DNSMASQ config
+# DNSMASQ constants
 # ==============================================================================
 
 ADDON_DHCP_SERVER_START_EPOCH="/data/startepoch"
@@ -11,6 +12,14 @@ DNSMASQ_LEASE_DATABASE="/data/dnsmasq.leases"
 
 # 5min is a reasonable threshold
 JUST_REBOOTED_THRESHOLD_SEC=300
+
+# ==============================================================================
+# FUNCTIONS
+# ==============================================================================
+
+function log_info() {
+    bashio::log.info "dnsmasq-init.sh: $@"
+}
 
 function ipvalid() {
   # Set up local variables
@@ -27,21 +36,21 @@ function ipvalid() {
 }
 
 function resolve_ntp_servers() {
-    NTP_SERVERS="$(jq --raw-output '.dhcp_network.ntp_servers[]' ${ADDON_CONFIG_RESOLVED} 2>/dev/null)"
+    NTP_SERVERS="$(jq --raw-output '.dhcp_server.ntp_servers[]' ${ADDON_CONFIG_RESOLVED} 2>/dev/null)"
     if [[ ! -z "${NTP_SERVERS}" ]]; then
-        bashio::log.info "NTP servers are ${NTP_SERVERS//$'\n'/,}"
+        log_info "NTP servers are ${NTP_SERVERS//$'\n'/,}"
 
         NTP_SERVERS_RESOLVED=""
         for srv in ${NTP_SERVERS}; do
             if ipvalid "$srv"; then
                 # no need to carry out any DNS resolution
-                echo "Using NTP IP $srv without any DNS resolution"
+                log_info "Using NTP IP $srv without any DNS resolution"
                 NTP_SERVERS_RESOLVED+="\"${srv}\","
             else
                 # run the DNS resolution and pick the first IP address
                 ip_addr="$(dig +short $srv | head -1)"
                 if [[ ! -z "${ip_addr}" ]]; then
-                    echo "DNS resolved $srv -> $ip_addr"
+                    log_info "DNS resolved $srv -> $ip_addr"
                     NTP_SERVERS_RESOLVED+="\"${ip_addr}\","
                 fi
             fi
@@ -52,54 +61,39 @@ function resolve_ntp_servers() {
             NTP_SERVERS_RESOLVED=${NTP_SERVERS_RESOLVED::-1}
 
             # add DNS-resolved IP addresses
-            jq --compact-output ".dhcp_network.ntp_resolved=[$NTP_SERVERS_RESOLVED]" \
+            jq --compact-output ".dhcp_server.ntp_resolved=[$NTP_SERVERS_RESOLVED]" \
                 ${ADDON_CONFIG_RESOLVED} >${ADDON_CONFIG_RESOLVED}.tmp
             mv ${ADDON_CONFIG_RESOLVED}.tmp ${ADDON_CONFIG_RESOLVED}
         fi
     fi
 }
 
-function get_dnsmasq_interface_ip_address() {
-    INTERFACE="$(jq --raw-output '.interface' ${ADDON_CONFIG} 2>/dev/null)"
-
-    # Get the IP address of the specified interface
-    INTERFACE_IP=$(ip -4 addr show "$INTERFACE" | grep -o 'inet [0-9]\{1,3\}\.[0-9]\{1,3\}\.[0-9]\{1,3\}\.[0-9]\{1,3\}' | awk '{print $2}')
-    if [[ -z "$INTERFACE_IP" ]]; then
-        echo "Error: Unable to determine IP address for interface $INTERFACE" >&2
-        exit 2
-    fi
-
-    echo "The IP address associated with the dnsmasq interface ${INTERFACE} is: ${INTERFACE_IP}"
-}
-
 function process_dns_servers() {
-    # fill the INTERFACE_IP global variable:
-    get_dnsmasq_interface_ip_address
-
-    DNS_SERVERS="$(jq --raw-output '.dhcp_network.dns_servers[]' ${ADDON_CONFIG} 2>/dev/null)"
+    DNS_SERVERS="$(jq --raw-output '.dhcp_server.dns_servers[]' ${ADDON_CONFIG} 2>/dev/null)"
     if [[ ! -z "${DNS_SERVERS}" ]]; then
-        bashio::log.info "DNS servers are ${DNS_SERVERS//$'\n'/,}"
+        log_info "DNS servers are ${DNS_SERVERS//$'\n'/,}"
 
         DNS_SERVERS_RESOLVED=""
         for srv in ${DNS_SERVERS}; do
             if ipvalid "$srv"; then
-                if [[ "$srv" == "0.0.0.0" ]]; then
-                    DNS_SERVERS_RESOLVED+="\"${INTERFACE_IP}\","
-                else
-                    DNS_SERVERS_RESOLVED+="\"${srv}\","
-                fi
+                # NOTE that dnsmasq supports the special address 0.0.0.0 which 
+                # is taken to mean "the address of the machine running dnsmasq".
+                # Since dnsmasq might be listening on multiple network interfaces, each
+                # with a different IP address, we need to delegate to dnsmasq the selection
+                # of the right IP address to advertise through DHCP
+                DNS_SERVERS_RESOLVED+="\"${srv}\","
             else
-                echo "Found invalid DNS server in DHCP network config: ${srv}. Skipping."
+                log_info "Found invalid DNS server in DHCP network config: ${srv}. Skipping."
             fi
         done
 
         if [[ ! -z "${DNS_SERVERS_RESOLVED}" ]]; then
             # pop last comma:
             DNS_SERVERS_RESOLVED=${DNS_SERVERS_RESOLVED::-1}
-            echo "List of processed DNS servers is: ${DNS_SERVERS_RESOLVED}"
+            log_info "List of processed DNS servers is: ${DNS_SERVERS_RESOLVED}"
 
             # add post-processed DNS servers
-            jq --compact-output ".dhcp_network.dns_servers_processed=[$DNS_SERVERS_RESOLVED]" \
+            jq --compact-output ".dhcp_server.dns_servers_processed=[$DNS_SERVERS_RESOLVED]" \
                 ${ADDON_CONFIG_RESOLVED} >${ADDON_CONFIG_RESOLVED}.tmp
             mv ${ADDON_CONFIG_RESOLVED}.tmp ${ADDON_CONFIG_RESOLVED}
         fi
@@ -109,7 +103,7 @@ function process_dns_servers() {
 function bump_dhcp_server_start_epoch() {
     updated_epoch="$(date +%s)"
     echo $updated_epoch > "$ADDON_DHCP_SERVER_START_EPOCH"
-    bashio::log.info "Updated DHCP start epoch is: $updated_epoch"
+    log_info "Updated DHCP start epoch is: $updated_epoch"
 }
 
 function reset_dhcp_leases_database_if_just_rebooted() {
@@ -118,7 +112,7 @@ function reset_dhcp_leases_database_if_just_rebooted() {
     uptime_seconds=$(awk '{print int($1)}' /proc/uptime)
 
     if [ "$uptime_seconds" -lt "$JUST_REBOOTED_THRESHOLD_SEC" ]; then
-        bashio::log.info "The HomeAssistant server has just been rebooted. Resetting DHCP lease database as requested in addon configuration."
+        log_info "The HomeAssistant server has just been rebooted. Resetting DHCP lease database as requested in addon configuration."
 
         # Get the current timestamp
         local timestamp
@@ -127,7 +121,7 @@ function reset_dhcp_leases_database_if_just_rebooted() {
         # the previuos database does not really get deleted, just moved in a file ignored by dnsmasq
         mv ${DNSMASQ_LEASE_DATABASE} ${DNSMASQ_LEASE_DATABASE}.${timestamp}
     else
-        bashio::log.info "The HomeAssistant server is up since ${uptime_seconds}secs. Skipping DHCP lease database reset."
+        log_info "The HomeAssistant server is up since ${uptime_seconds}secs. Skipping DHCP lease database reset."
     fi
 }
 
@@ -136,36 +130,39 @@ function reset_dhcp_leases_database_if_just_rebooted() {
 # MAIN
 #
 
-bashio::log.info "Starting dnsmasq configuration..."
+log_info "Starting dnsmasq configuration..."
 
 should_reset_on_reboot=$(bashio::config 'dhcp_server.reset_dhcp_lease_database_on_reboot')
 if [[ "$should_reset_on_reboot" = "null" ]]; then
     should_reset_on_reboot=false
 fi
-bashio::log.info The setting reset_dhcp_lease_database_on_reboot is ${should_reset_on_reboot}"..."
+log_info The setting reset_dhcp_lease_database_on_reboot is ${should_reset_on_reboot}"..."
 if $should_reset_on_reboot ; then
     reset_dhcp_leases_database_if_just_rebooted
 fi
 
-bashio::log.info "Advancing the DHCP server start epoch..."
+log_info "Advancing the DHCP server start epoch..."
 bump_dhcp_server_start_epoch
 
-# by deafult the resolved config is equal to the original config
+# by default the resolved config is equal to the original config
 cp ${ADDON_CONFIG} ${ADDON_CONFIG_RESOLVED}
 
 # do some processing:
-bashio::log.info "Resolving NTP hostnames eventually provided..."
+log_info "Resolving NTP hostnames eventually provided..."
 resolve_ntp_servers
-bashio::log.info "Processing DHCP DNS server list..."
+log_info "Processing DHCP DNS server list..."
 process_dns_servers
 
-bashio::log.info "Configuring dnsmasq..."
+log_info "Configuring dnsmasq..."
 tempio \
     -conf ${ADDON_CONFIG_RESOLVED} \
     -template /usr/share/tempio/dnsmasq.config \
     -out "${DNSMASQ_CONFIG}"
 
-bashio::log.info "Full dnsmasq config:"
+# make the config file a bit more compact by removing empty lines
+sed -i '/^$/d' $DNSMASQ_CONFIG
+
+log_info "Full dnsmasq config:"
 cat -n $DNSMASQ_CONFIG
 
-bashio::log.info "Successfully completed dnsmasq configuration."
+log_info "Successfully completed dnsmasq configuration."